+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Minor Features
 
 - Add a new method `flow.get_tasks()` for easily filtering flow tasks by attribute [#242](https://github.com/PrefectHQ/prefect/pull/242)
+- Add new `JinjaTemplateTask` for easily rendering jinja templates [#200](https://github.com/PrefectHQ/prefect/issues/200)
+- Add new `PAUSE` signal for halting task execution [#246](https://github.com/PrefectHQ/prefect/pull/246)
+- Add ability to timeout task execution for all executors except `DaskExecutor(processes=True)` [#240](https://github.com/PrefectHQ/prefect/issues/240)
 
 ### Fixes